Moved the header buttons specific to the data browser in the databrowser/ folder Moved the ButtonsContainer component in ui-elements/Container.jsx Used the addTrailingSlash method from the utilities
